backend/chatbot/requirements.txt
python-dotenv==1.0.1
openai==1.12.0
httpx>=0.24.1 